fix: resolve Clang 17 build warnings and errors
- byte_order.hpp: add explicit static_cast for uint16_t to fix implicit conversion warnings
- uds_types.hpp: add explicit casts for bit operations returning to smaller types
- doip.cpp: replace std::format with std::ostringstream for Clang libc++ compatibility
- rtps_decoder.cpp, gptp_decoder.cpp, uds_decoder.cpp: mark unused helper functions
- matchers.hpp: add defensive null check for GCC null-dereference warning
- generators.hpp, live_assert.hpp, record_replay.hpp: add missing includes
- test_decode_pipeline.cpp: add ASSERT_NE checks after get_layer calls

All tests pass with both GCC and Clang compilers.

+++ include/wadjet/core/byte_order.hpp
@@ -12,8 +12,9 @@ namespace wadjet {
 
 /// @brief Read a 16-bit value in network byte order (big-endian)
 [[nodiscard]] inline constexpr std::uint16_t read_be16(const std::byte* data) {
-    return (static_cast<std::uint16_t>(static_cast<std::uint8_t>(data[0])) << 8) |
-           static_cast<std::uint16_t>(static_cast<std::uint8_t>(data[1]));
+    return static_cast<std::uint16_t>(
+        (static_cast<std::uint16_t>(static_cast<std::uint8_t>(data[0])) << 8) |
+        static_cast<std::uint16_t>(static_cast<std::uint8_t>(data[1])));
 }
 
 /// @brief Read a 32-bit value in network byte order (big-endian)
@@ -52,8 +53,9 @@ inline constexpr void write_be64(std::byte* data, std::uint64_t value) {
 
 /// @brief Read a 16-bit value in little-endian order
 [[nodiscard]] inline constexpr std::uint16_t read_le16(const std::byte* data) {
-    return static_cast<std::uint16_t>(static_cast<std::uint8_t>(data[0])) |
-           (static_cast<std::uint16_t>(static_cast<std::uint8_t>(data[1])) << 8);
+    return static_cast<std::uint16_t>(
+        static_cast<std::uint16_t>(static_cast<std::uint8_t>(data[0])) |
+        (static_cast<std::uint16_t>(static_cast<std::uint8_t>(data[1])) << 8));
 }
 
 /// @brief Read a 32-bit value in little-endian order
